requirements.txt
opencv-python
easyocr
numpy
pandas
gspread
google-auth
google-auth-oauthlib
requests